pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raidhub-discord"
version = "0.1.0"
description = "RaidHub Discord ingress and command service"
requires-python = ">=3.14"
dependencies = [
  "discord-interactions",
  "fastapi",
  "uvicorn",
  "pynacl",
  "httpx",
  "python-dotenv",
  "PyJWT",
  "prometheus-client",
  "sentry-sdk[fastapi]",
]

[project.scripts]
sync-discord-commands = "src.cli_sync_commands:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src"]

[project.optional-dependencies]
dev = [
  "ruff",
  "mypy"
]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.14"
strict = false
ignore_missing_imports = true
